Integrate spin and pincpu into pulse
Trigger pulses on defined CPU set. Threads are dispatched in a round-robin way.

Example:

  wk-pulse --cpuset 0,0,1 --thread 4

  thread 0 -> CPU 0
  thread 1 -> CPU 0
  thread 2 -> CPU 1
  thread 3 -> CPU 0

Thus, thread 0,1,3 will wake-up at the same time and fight for the CPU for the
specified period of time. The period is fixed, if contention is very high, then
there is no guarantee a thread will run at all.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>